--- a/pkg/kv/kvserver/rangefeed/event_size.go
+++ b/pkg/kv/kvserver/rangefeed/event_size.go
@@ -33,8 +33,7 @@ const (
 	mvccAbortIntentOp  = int64(unsafe.Sizeof(enginepb.MVCCAbortIntentOp{}))
 	mvccAbortTxnOp     = int64(unsafe.Sizeof(enginepb.MVCCAbortTxnOp{}))
 
-	eventPtrOverhead = int64(unsafe.Sizeof(&event{}))
-	eventOverhead    = int64(unsafe.Sizeof(&event{})) + int64(unsafe.Sizeof(event{}))
+	eventOverhead = int64(unsafe.Sizeof(&event{})) + int64(unsafe.Sizeof(event{}))
 
 	sstEventOverhead  = int64(unsafe.Sizeof(sstEvent{}))
 	syncEventOverhead = int64(unsafe.Sizeof(syncEvent{}))
@@ -184,19 +183,9 @@ func abortTxnOpMemUsage(txnID uuid.UUID) int64 {
 	return currMemUsage
 }
 
-// eventOverhead accounts for the base struct of event{} which only included
-// pointer to syncEvent. currMemUsage accounts the base struct memory of
-// syncEvent{} and its underlying memory.
-func (sync syncEvent) currMemUsage() int64 {
-	// syncEvent has a channel and a pointer to testRegCatchupSpan.
-	// testRegCatchupSpan is never set in production. Channel sent is always
-	// struct{}, so the memory has already been accounted in syncEventOverhead.
-	return syncEventOverhead
-}
-
 // currMemUsage returns the current memory usage of the opsEvent including base
 // structs overhead and underlying memory usage.
-func (ops opsEvent) currMemUsage() int64 {
+func opsCurrMemUsage(ops opsEvent) int64 {
 	// currMemUsage: eventOverhead already accounts for slice overhead in
 	// opsEvent, []enginepb.MVCCLogicalOp. For each cap(ops), the underlying
 	// memory include a MVCCLogicalOp overhead.
@@ -263,11 +252,7 @@ func (e *event) String() string {
 
 // MemUsage estimates the total memory usage of the event, including its
 // underlying data. The memory usage is estimated in bytes.
-func (e *event) MemUsage() int64 {
-	if e == nil {
-		// For nil event, only eventPtrOverhead is accounted.
-		return eventPtrOverhead
-	}
+func MemUsage(e event) int64 {
 	// currMemUsage: pointer to e is passed to p.eventC channel. Each e pointer is
 	// &event{}, and each pointer points at an underlying event{}.
 	switch {
@@ -275,7 +260,7 @@ func (e *event) MemUsage() int64 {
 		// For logical ops events, current memory usage is usually larger than
 		// rangefeed events. Note that we assume no checkpoint events are caused by
 		// ops since they are pretty rare to avoid the complexity.
-		return eventOverhead + e.ops.currMemUsage()
+		return eventOverhead + opsCurrMemUsage(e.ops)
 	case !e.ct.IsEmpty():
 		// For ct event, rangefeed checkpoint event usually takes more memory than
 		// current memory usage. Note that we assume checkpoint event will happen.
@@ -294,7 +279,7 @@ func (e *event) MemUsage() int64 {
 		return rangefeedSSTTableOpMemUsage(e.sst.data, e.sst.span.Key, e.sst.span.EndKey)
 	case e.sync != nil:
 		// For sync event, no rangefeed events will be published.
-		return eventOverhead + e.sync.currMemUsage()
+		return eventOverhead + syncEventOverhead
 	default:
 		log.Fatalf(context.Background(), "missing event variant: %+v", e)
 	}