--- a/pkg/kv/kvserver/rangefeed/scheduled_processor.go
+++ b/pkg/kv/kvserver/rangefeed/scheduled_processor.go
@@ -448,7 +448,7 @@ func (p *ScheduledProcessor) enqueueEventInternal(
 	// inserting value into channel.
 	var alloc *SharedBudgetAllocation
 	if p.MemBudget != nil {
-		size := e.MemUsage()
+		size := MemUsage(e)
 		if size > 0 {
 			var err error
 			// First we will try non-blocking fast path to allocate memory budget.