@wenyihu6 2024-05-15
Yea, this is just an improvement but does't fully solve the regression. The current status in https://github.com/cockroachdb/cockroach/issues/123076#issuecomment-2083275524 has the new benchmark against a commit before the work.





<!-- Sent from Reviewable.io -->